next

+++ src/test/java/ru/alfastrah/embedi/tick/models/TickQuoteTestData.java
@@ -1,5 +1,6 @@
 package ru.alfastrah.embedi.tick.models;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.stream.IntStream;
 
@@ -11,7 +12,7 @@ public class TickQuoteTestData {
         tickQuote.setAddress(AddressTestData.makeFake(aSeed));
         tickQuote.setEndDate(LocalDateTime.of(2022, 11, 14, 15, 50).plusDays(aSeed));
         tickQuote.setStartDate(LocalDateTime.of(2021, 11, 14, 15, 50).plusDays(aSeed));
-        tickQuote.setPremium(100.50 + aSeed);
+        tickQuote.setPremium(BigDecimal.valueOf(100.50 + aSeed));
         tickQuote.setInsurer(PersonTestData.makeFake(aSeed));
         tickQuote.setInsuredPersons(
                 IntStream.range(0, 5).mapToObj(i -> PersonTestData.makeFake(aSeed + i)).toList());
